requirements.txt
openai>=1.0.0
pandas>=1.0.0
Pillow>=9.0.0
chardet>=5.0.0
python-dotenv>=1.0.0
